Solve day 5

+++ src/day_01.rs
@@ -9,7 +9,7 @@ use nom::{
     IResult,
 };
 
-pub fn solve() -> (u32, u32) {
+pub fn solve() -> (u64, u64) {
     let input = include_str!("../inputs/day_01");
 
     let mut sum: u32 = 0;
@@ -27,7 +27,7 @@ pub fn solve() -> (u32, u32) {
         sum2 += msd * 10 + lsd;
     });
 
-    (sum, sum2)
+    (sum as u64, sum2 as u64)
 }
 
 fn find_number(s: &str, reverse: bool) -> u32 {
